Add bind_server resource
This adds a simple bind_server resource to add server stanzas to the
named configuration.

I've modelled this directly after the current functionality to add
server options via an attribute. Given the rather limited amount of
things you can add to this stanza it would be good in future to model
those more closely so we can check the parameters a bit more carefully.