Support chroot environment on ubuntu 18.04
Previously chroot on ubuntu didn't work because of a conflict with the
default Ubuntu AppArmor profile for named.

This commit fixes that by including a local apparmor configuration which
contains the standard chroot environment. This file is only installed
and loaded in the event the cookbook is running with the chroot flag set
to true in the options, and the OS is Ubuntu 18.04 or later.

As part of this, we've also fixed the Debian chroot environment where
BIND was defaulting to /run within the chroot for the pid file and
session key file. These are now redirected to the /var/run/named
directory within the chroot environment.

See #52