Restart bind_service immediately when using delayed_action :create
I encountered an issue where the templates defined in bind_config were not
restarting bind after being created. It seems the issue is because those
resources are using ``delayed_action :create`` and with the ``:delayed`` timer.
The delayed restart notifiations seem to happen **before** the template gets
created. This results in a bind service not having the configuration I'm
expecting it to have. To work around this, this should be set to :immediately.

Signed-off-by: Lance Albertson <lance@osuosl.org>